requirements.txt
Flask==1.1.1
gunicorn==19.9.0
flask-cors==3.0.8
#Flask-PyMongo==2.3.0
python-jwt==3.2.4
#JWT dependencies
asn1crypto==0.24.0 
cffi==1.12.3
cryptography==2.7 
jwcrypto==0.6.0 
pycparser==2.19 
PyJWT==1.7.1
PyYAML==5.1.2
requests==2.22.0
